⏪️ revert(api): Revert "♻️ refactor(api/services): update import paths to use models directory"
