+++ apps/api/src/app/auth/[client]/logout.ts
@@ -4,10 +4,10 @@ import { SESSION_COOKIE_NAME } from "../../../common/constants";
 import { clientSchema } from "../../../common/schema";
 import { DrizzleService } from "../../../infrastructure/drizzle";
 import { SessionRepository } from "../../../interface-adapter/repositories/session";
-import { CookieService } from "../../../models/services/cookie";
-import { SessionTokenService } from "../../../models/services/session-token";
 import { authGuard } from "../../../modules/auth-guard";
 import { ElysiaWithEnv } from "../../../modules/elysia-with-env";
+import { CookieService } from "../../../services/cookie";
+import { SessionTokenService } from "../../../services/session-token";
 
 const cookieSchemaObject = {
 	[SESSION_COOKIE_NAME]: t.Optional(t.String()),
