+++ apps/api/src/application/use-cases/auth/logout.usecase.ts
@@ -1,5 +1,5 @@
 import type { ISessionRepository } from "../../../interface-adapter/repositories/session";
-import type { ISessionTokenService } from "../../../models/services/session-token";
+import type { ISessionTokenService } from "../../../services/session-token";
 import type { ILogoutUseCase } from "./interfaces/logout.usecase.interface";
 
 export class LogoutUseCase implements ILogoutUseCase {
