+++ apps/api/src/services/cookie/interfaces/cookie.service.interface.ts
