+++ apps/api/src/services/password/interfaces/password.service.interface.ts
