+++ apps/api/src/services/session-token/interfaces/session-token.service.interface.ts
